chore(checkout): Deprecate risk functions
