--- packages/checkout/sdk/src/riskAssessment/riskAssessment.ts
@@ -1,45 +1,0 @@
-import axios from 'axios';
-import { IMMUTABLE_API_BASE_URL } from '../env';
-import { RiskAssessmentConfig } from '../types';
-import { CheckoutConfiguration } from '../config';
-import { AssessmentResult, RiskAssessmentResponse } from './common';
-
-export const fetchRiskAssessment = async (
-  addresses: string[],
-  config: CheckoutConfiguration,
-): Promise<AssessmentResult> => {
-  const result = Object.fromEntries(
-    addresses.map((address) => [address.toLowerCase(), { sanctioned: false }]),
-  );
-
-  const riskConfig = (await config.remote.getConfig('riskAssessment')) as
-    | RiskAssessmentConfig
-    | undefined;
-
-  if (!riskConfig?.enabled) {
-    return result;
-  }
-
-  try {
-    const riskLevels = riskConfig?.levels.map((l) => l.toLowerCase()) ?? [];
-
-    const response = await axios.post<RiskAssessmentResponse[]>(
-      `${IMMUTABLE_API_BASE_URL[config.environment]}/v1/sanctions/check`,
-      {
-        addresses,
-      },
-    );
-
-    for (const assessment of response.data) {
-      result[assessment.address.toLowerCase()].sanctioned = riskLevels.includes(
-        assessment.risk.toLowerCase(),
-      );
-    }
-
-    return result;
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Error fetching risk assessment', error);
-    return result;
-  }
-};
