+++ packages/checkout/sdk/tsconfig.json
@@ -5,6 +5,6 @@
     "rootDirs": ["src"],
     "customConditions": ["development"]
   },
-  "include": ["src"],
+  "include": ["src", "../widgets-lib/src/lib/riskAssessment"],
   "exclude": ["node_modules", "dist", "test"]
 }
