+++ packages/checkout/widgets-lib/src/lib/riskAssessment/index.ts
@@ -1,3 +1,2 @@
 export * from './riskAssessment';
-export * from './riskAssessmentV2';
 export * from './common';
